+++ wire/internal/wire/testdata/UnexportedStruct/foo/foo.go
@@ -17,5 +17,5 @@ package main
 import "fmt"
 
 func main() {
-	fmt.Println(injectedMessage())
+	fmt.Printf("%v\n", injectedBar())
 }
