+++ wire/internal/wire/testdata/UnexportedStruct/pkg
@@ -1,0 +1,1 @@
+example.com/foo
