+++ wire/internal/wire/testdata/UnexportedStruct/want/wire_errs.txt
@@ -1,0 +1,1 @@
+/wire_gopath/src/example.com/foo/wire.go:x:y: foo not exported by package bar
