+++ wire/internal/wire/testdata/UnexportedValue/want/wire_errs.txt
@@ -1,1 +1,1 @@
-/wire_gopath/src/example.com/foo/wire.go:x:y: bar not exported by package bar
+/wire_gopath/src/example.com/foo/wire.go:x:y: inject injectedMessage: value string can't be used: uses unexported identifier privateMsg
